--- src/initialFolders/addPost.html
@@ -1,35 +1,0 @@
-<div class="content">
-  <h2>title</h2>
-
-  <form>
-    <div class="field" data-cy="NameField">
-      <label class="label" for="comment-author-name-title">Title</label>
-      <div class="control has-icons-left has-icons-right">
-        <input
-          type="text"
-          name="title"
-          id="comment-author-name-title"
-          placeholder="Post title"
-          class="input"
-          value=""
-        /><span class="icon is-small is-left"><i class="fas fa-user fa-heading"></i></span>
-      </div>
-    </div>
-
-    <div class="field" data-cy="BodyField">
-      <label class="label" for="comment-body">Write Post Body</label>
-      <div class="control">
-        <textarea id="comment-body" name="body" placeholder="Post body" class="textarea"></textarea>
-      </div>
-    </div>
-
-    <div class="field is-grouped">
-      <div class="control">
-        <button type="submit" class="button is-link">Save</button>
-      </div>
-      <div class="control">
-        <button type="reset" class="button is-link is-light">Cancel</button>
-      </div>
-    </div>
-  </form>
-</div>
