--- src/initialFolders/comment.html
@@ -1,7 +1,0 @@
-<article class="message is-small">
-  <div class="message-header">
-    <a href="{`mailto:${email}`}"> {{ name }} </a>
-    <button type="button" class="delete is-small" aria-label="delete">delete button</button>
-  </div>
-  <div class="message-body">{{ el body }}</div>
-</article>
