--- src/initialFolders/inputField.html
@@ -1,21 +1,0 @@
-<div class="field" data-cy="NameField">
-  <label class="label" htmlFor="{`comment-author-name-${name}`}"> title </label>
-  <div class="control has-icons-left has-icons-right">
-    <input
-      type="text"
-      name="{name}"
-      id="{`comment-author-name-${name}`}"
-      placeholder="{placeholder}"
-      class="input is-danger"
-    />
-    <span class="icon is-small is-left">
-      <i class="fas fa-user"></i>
-    </span>
-
-    <span class="icon is-small is-right has-text-danger" data-cy="ErrorIcon">
-      <i class="fas fa-exclamation-triangle"></i>
-    </span>
-  </div>
-
-  <p class="help is-danger" data-cy="ErrorMessage">error text</p>
-</div>
