--- src/initialFolders/noCommentsYet.html
@@ -1,3 +1,0 @@
-<div class="block">
-  <p class="title is-4">No comments yet</p>
-</div>
