--- src/initialFolders/postLoader.html
@@ -1,3 +1,0 @@
-<div class="is-flex is-justify-content-center is-align-items-center mt-2">
-  <!-- Loader component -->
-</div>
