--- src/initialFolders/postPreview.html
@@ -1,14 +1,0 @@
-<div class="block">
-  <div class="is-flex is-justify-content-space-between is-align-items-center">
-    <h2>#post id: post title</h2>
-    <div class="is-flex">
-      <span class="icon is-small is-right is-clickable">
-        <i class="fas fa-pen-to-square"></i>
-      </span>
-      <span class="icon is-small is-right has-text-danger is-clickable ml-3">
-        <i class="fas fa-trash"></i>
-      </span>
-    </div>
-  </div>
-  <p data-cy="PostBody">post body</p>
-</div>
